+++ src/lib/voxel/filter-floaters.ts
@@ -4,14 +4,11 @@ import {
     buildBlockGridParams
 } from './filter-pipeline';
 import {
-    alignGridBounds
-} from './sparse-octree';
-import {
     buildBlockLookup,
     isCenterInOccupiedVoxel,
     gaussianContributesToVoxels
 } from './voxel-query';
-import { voxelizeToBuffer } from './voxelize';
+import { alignGridBounds, voxelizeToBuffer } from './voxelize';
 import { DataTable } from '../data-table';
 import type { DeviceCreator } from '../types';
 import { logger } from '../utils';
