+++ src/lib/voxel/voxelize.ts
@@ -1,6 +1,8 @@
+import { Vec3 } from 'playcanvas';
+
 import { BlockMaskBuffer } from './block-mask-buffer';
 import { xyzToMorton } from './morton';
-import type { Bounds } from './sparse-octree';
+import type { Bounds } from '../data-table';
 import {
     GpuVoxelization,
     type BatchSpec,
@@ -206,4 +208,36 @@ const voxelizeToBuffer = async (
     return buffer;
 };
 
-export { voxelizeToBuffer };
+/**
+ * Align bounds to 4x4x4 block boundaries.
+ *
+ * @param minX - Scene minimum X
+ * @param minY - Scene minimum Y
+ * @param minZ - Scene minimum Z
+ * @param maxX - Scene maximum X
+ * @param maxY - Scene maximum Y
+ * @param maxZ - Scene maximum Z
+ * @param voxelResolution - Size of each voxel
+ * @returns Aligned bounds
+ */
+function alignGridBounds(
+    minX: number, minY: number, minZ: number,
+    maxX: number, maxY: number, maxZ: number,
+    voxelResolution: number
+): Bounds {
+    const blockSize = 4 * voxelResolution;
+    return {
+        min: new Vec3(
+            Math.floor(minX / blockSize) * blockSize,
+            Math.floor(minY / blockSize) * blockSize,
+            Math.floor(minZ / blockSize) * blockSize
+        ),
+        max: new Vec3(
+            Math.ceil(maxX / blockSize) * blockSize,
+            Math.ceil(maxY / blockSize) * blockSize,
+            Math.ceil(maxZ / blockSize) * blockSize
+        )
+    };
+}
+
+export { voxelizeToBuffer, alignGridBounds };
